+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.2.0] - 2025-01-11
+### Added
+- New command `lume local` to use a local repository of Lume
+  - `lume local --save`: Save the current directory as the path Lume.
+  - `lume local`: Apply the Lume path into `lume/` import map entry.
+  - `lume local --remove`: Remove the previously saved path of Lume.
+
 ## [2.1.1] - 2024-05-22
 ### Fixed
 - Added the `--global` flag needed for incoming Deno 2.
@@ -42,6 +49,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.0.0] - 2023-01-06
 First version
 
+[2.2.0]: https://github.com/lumeland/cli/compare/v2.1.1...v2.2.0
 [2.1.1]: https://github.com/lumeland/cli/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/lumeland/cli/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/lumeland/cli/compare/v1.0.3...v2.0.0